+++ crates/wdl-analysis/src/handlers/rename.rs
@@ -37,13 +37,13 @@ pub fn rename(
         bail!("name `{new_name}` is not a valid WDL identifier");
     }
 
-    let locations = handlers::find_all_references(graph, document_uri, position, encoding, true)?;
-    if locations.is_empty() {
+    let references = handlers::find_all_references(graph, document_uri, position, encoding, true)?;
+    if references.is_empty() {
         return Ok(None);
     }
 
     let mut changes: HashMap<Url, Vec<TextEdit>> = HashMap::new();
-    for location in locations {
+    for location in references.locations() {
         let text_edit = TextEdit {
             range: location.range,
             new_text: new_name.clone(),